+++ docs/api/puppeteer.createbrowserfetcher.md
@@ -9,7 +9,7 @@ sidebar_label: createBrowserFetcher
 ```typescript
 createBrowserFetcher: (
   options: Partial<
-    import('puppeteer-core/internal/node/node.js').BrowserFetcherOptions
+    import('puppeteer-core/internal/puppeteer-core.js').BrowserFetcherOptions
   >
-) => import('puppeteer-core/internal/node/node.js').BrowserFetcher;
+) => import('puppeteer-core/internal/puppeteer-core.js').BrowserFetcher;
 ```
