+++ packages/puppeteer-core/src/common/common.ts
@@ -26,9 +26,11 @@ export * from './ConnectionTransport.js';
 export * from './ConsoleMessage.js';
 export * from './Coverage.js';
 export * from './Debug.js';
+export * from './Device.js';
 export * from './Dialog.js';
 export * from './ElementHandle.js';
 export * from './EmulationManager.js';
+export * from './Errors.js';
 export * from './EventEmitter.js';
 export * from './ExecutionContext.js';
 export * from './fetch.js';
@@ -41,6 +43,7 @@ export * from './HTTPRequest.js';
 export * from './HTTPResponse.js';
 export * from './Input.js';
 export * from './IsolatedWorld.js';
+export * from './IsolatedWorlds.js';
 export * from './JSHandle.js';
 export * from './LazyArg.js';
 export * from './LifecycleWatcher.js';
@@ -49,6 +52,7 @@ export * from './NetworkManager.js';
 export * from './NodeWebSocketTransport.js';
 export * from './Page.js';
 export * from './PDFOptions.js';
+export * from './PredefinedNetworkConditions.js';
 export * from './Product.js';
 export * from './Puppeteer.js';
 export * from './PuppeteerViewport.js';
